added getCellRects

+++ ClientSide/GraphicsFunctions/Maze/node.h
@@ -4,8 +4,8 @@
 class node {
 	private:
 	public:
-		node(std::vector<int> connectionsP, sf::Vector2i posP);
+		node(std::vector<int> connectionsP, sf::Vector2f posP);
 		node();
-		sf::Vector2i pos;
-		std::vector<sf::Vector2i> connections;//list of positions in maze that this node is connected to
+		sf::Vector2f pos;
+		std::vector<sf::Vector2f> connections;//list of positions in maze that this node is connected to
 };
